chore(Mail Thread): move out placeholder
